Create validate-binary-search-tree.cpp
Transformation of in-order traversal. No need for extra O(n) space for the entire sequence of traversal.
Use a pointer to point to last node traversed and compare to current pointer.
First traverse to the left child.
Then update the pointer to point to current node.
At last, traverse to the right child.